requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.10.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.73.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.3
pydantic-core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
uritemplate==4.1.1
urllib3==2.4.0
websockets==15.0.1
